Run with default logging. Sum up on compare errors. Add quiet flag

+++ tests/local.sh
@@ -16,14 +16,20 @@ fi
 
 : ${PROG:=../csync2}
 
+: ${DEBUG:=v}
+
 COUNT=0
 echo "Running command $COMMAND" 
+RES=0
 
 function testing {
     RESULT=$1
     OLD_RESULT=${RESULT}.res
     if [ -f ${OLD_RESULT} ] ; then
-	diff $RESULT $OLD_RESULT | tee ${RESULT}.diff
+	diff $RESULT $OLD_RESULT > ${RESULT}.diff
+	rc=$?
+	RES=$((RES+rc))
+	cat ${RESULT}.diff
     else
 	cp $RESULT $OLD_RESULT
     fi
@@ -47,15 +53,17 @@ function cmd {
     fi
     echo cmd $CMD \"$2\" $HOST > ${TESTNAME}/${COUNT}.log 
     if [ "$GDB" != "" ] ; then 
-	$GDB $PROG -P peer -p 30860 -K csync2_$HOST.cfg -N $HOST -${CMD}r$DEBUG test
+	$GDB $PROG -q -P peer -p 30860 -K csync2_$HOST.cfg -N $HOST -${CMD}r$DEBUG test
     else
-	$PROG -P peer -p 30860 -K csync2_$HOST.cfg -N $HOST -${CMD}r$DEBUG test >> ${TESTNAME}/${COUNT}.log 2>&1
+	$PROG -q -P peer -p 30860 -K csync2_$HOST.cfg -N $HOST -${CMD}r$DEBUG test >> ${TESTNAME}/${COUNT}.log 2>&1
     fi
     testing ${TESTNAME}/${COUNT}.log
-    echo "select filename from file order by filename; select peername,filename,operation,other from dirty order by filename;" | mysql -t -u csync2_$HOST -pcsync2_$HOST csync2_$HOST > ${TESTNAME}/${COUNT}.mysql 2> /dev/null
+    echo "select filename from file order by filename; select peername,filename,operation,other from dirty order by filename, peername;" | mysql -t -u csync2_$HOST -pcsync2_$HOST csync2_$HOST > ${TESTNAME}/${COUNT}.mysql 2> /dev/null
     testing ${TESTNAME}/${COUNT}.mysql
-    rsync --delete -nHav test/local/ peer:`pwd`/test/peer/ |grep -v "bytes/sec" |grep -v "(DRY RUN)" |grep -v "sending incremental" > ${TESTNAME}/${COUNT}.rsync
-    testing ${TESTNAME}/${COUNT}.rsync
+    if [ -d "test/local" ] ; then 
+	rsync --delete -nHav test/local/ peer:`pwd`/test/peer/ |grep -v "bytes/sec" |grep -v "(DRY RUN)" |grep -v "sending incremental" > ${TESTNAME}/${COUNT}.rsync
+	testing ${TESTNAME}/${COUNT}.rsync
+    fi
     echo "${COUNT}. END $CMD ${DESC}" 
     let COUNT=$COUNT+1
     ${PAUSE}
@@ -79,17 +87,17 @@ function daemon {
     CMD="$1"
     echo $NAME $PEER
     if [ "$CMD" == "d" ] ; then 
-	${PROG} -K csync2_$PEER.cfg -N $PEER -z $NAME -iiii$DEBUG -p 30860 > $TESTNAME/daemon.log  2>&1 &
+	${PROG} -q -K csync2_$PEER.cfg -N $PEER -z $NAME -iiii$DEBUG -p 30860 > $TESTNAME/daemon.log  2>&1 &
 	echo "$!" > daemon.pid
     elif [ "$CMD" == "i" ] ; then 
-	$GDB ${PROG} -K csync2_$NAME.cfg -N $NAME -z $PEER -iiii$DEBUG -p 30860 > $TESTNAME/daemon.log  2>&1
+	$GDB ${PROG} -q -K csync2_$NAME.cfg -N $NAME -z $PEER -iiii$DEBUG -p 30860 > $TESTNAME/daemon.log  2>&1
 	echo "$!" > daemon.pid
 	sleep 1
     elif [ "$CMD" == "once" ] ; then 
-	${PROG} -K csync2_$NAME.cfg -N $NAME -z $PEER -iii$DEBUG -p 30860 >> daemon_${NAME}.log 2>&1 & 
+	${PROG} -q -K csync2_$NAME.cfg -N $NAME -z $PEER -iii$DEBUG -p 30860 >> daemon_${NAME}.log 2>&1 & 
     elif [ "$CMD" == "clean_once" ] ; then 
 	clean peer
-	${PROG} -K csync2_$NAME.cfg -N $NAME -z $PEER -iii$DEBUG -p 30860 & 
+	${PROG} -q -K csync2_$NAME.cfg -N $NAME -z $PEER -iii$DEBUG -p 30860 & 
     fi    
 }
 
@@ -129,3 +137,5 @@ for d in $* ; do
 	fi 
     fi
 done
+echo "Result $RES"
+exit $RES
